requirements.txt
asgiref==3.4.1
Django==3.2.12
django-environ==0.8.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.4.0
pkg_resources==0.0.0
psycopg2-binary==2.9.3
PyJWT==2.3.0
pytz==2022.1
sqlparse==0.4.2
typing_extensions==4.1.1
